perf(seo): batch bulk worker metadata reads
