+++ crates/rustok-seo/src/services/bulk_read_model.rs
@@ -4,65 +4,64 @@ use rustok_api::TenantContext;
 use rustok_content::resolve_by_locale_with_fallback;
 use rustok_seo_targets::{SeoTargetBulkListRequest, SeoTargetSlug};
 use sea_orm::{ColumnTrait, EntityTrait, QueryFilter, QueryOrder};
+use serde_json::Value;
 use uuid::Uuid;
 
 use crate::dto::{SeoBulkItem, SeoBulkListInput, SeoBulkPage, SeoBulkSource, SeoModuleSettings};
 use crate::entities::{self as seo_meta, meta_translation};
 use crate::{SeoError, SeoResult};
 
-use super::super::templates::render_generated_record;
-use super::super::{trimmed_option, LoadedMeta, SeoService, TargetState};
+use super::robots::first_open_graph_image_url;
+use super::templates::render_generated_record;
+use super::{trimmed_option, LoadedMeta, SeoService, TargetState};
 
 const MAX_BULK_PAGE_SIZE: i32 = 100;
 const BULK_META_BATCH_SIZE: usize = 256;
 
 #[derive(Debug, Clone)]
-struct BatchedBulkListFilter {
-    target_kind: SeoTargetSlug,
-    locale: String,
-    query: Option<String>,
-    source: SeoBulkSource,
-    page: i32,
-    per_page: i32,
+pub(super) struct BulkReadFilter {
+    pub target_kind: SeoTargetSlug,
+    pub locale: String,
+    pub query: Option<String>,
+    pub source: SeoBulkSource,
 }
 
 #[derive(Debug, Clone)]
-struct BatchedBulkSummary {
-    target_id: Uuid,
-    label: String,
-    route: String,
+pub(super) struct BulkReadProjection {
+    pub effective_locale: String,
+    pub source: SeoBulkSource,
+    pub title: Option<String>,
+    pub description: Option<String>,
+    pub keywords: Option<String>,
+    pub canonical_url: Option<String>,
+    pub og_title: Option<String>,
+    pub og_description: Option<String>,
+    pub og_image: Option<String>,
+    pub structured_data: Option<Value>,
+    pub noindex: bool,
+    pub nofollow: bool,
 }
 
 #[derive(Debug, Clone)]
-struct BatchedBulkMeta {
-    effective_locale: String,
-    source: SeoBulkSource,
-    title: Option<String>,
-    description: Option<String>,
-    canonical_url: Option<String>,
-    noindex: bool,
-    nofollow: bool,
-}
-
-#[derive(Debug, Clone)]
-struct BatchedBulkRow {
-    summary: BatchedBulkSummary,
-    meta: BatchedBulkMeta,
+pub(super) struct BulkReadRow {
+    pub target_id: Uuid,
+    pub label: String,
+    pub route: String,
+    pub projection: BulkReadProjection,
 }
 
 impl SeoService {
-    pub(super) async fn list_bulk_items_batched(
+    pub(super) async fn collect_bulk_read_rows(
         &self,
         tenant: &TenantContext,
-        input: SeoBulkListInput,
-    ) -> SeoResult<SeoBulkPage> {
-        let filter = normalize_batched_bulk_list_input(input, tenant.default_locale.as_str())?;
+        filter: &BulkReadFilter,
+    ) -> SeoResult<Vec<BulkReadRow>> {
         if !self.is_enabled(tenant.id).await? {
-            return Ok(empty_bulk_page(&filter));
+            return Ok(Vec::new());
         }
 
         let Some(provider) = self.registry.get(&filter.target_kind) else {
-            return Ok(empty_bulk_page(&filter));
+            return Ok(Vec::new());
         };
         let summaries = provider
             .list_bulk_summaries(
@@ -79,15 +78,7 @@ impl SeoService {
                     "SEO target provider `{}` failed to collect bulk summaries: {error}",
                     filter.target_kind.as_str()
                 ))
-            })?
-            .into_iter()
-            .map(|summary| BatchedBulkSummary {
-                target_id: summary.target_id,
-                label: summary.label,
-                route: summary.route,
-            })
-            .collect::<Vec<_>>();
-
+            })?;
         let target_ids = summaries
             .iter()
             .map(|summary| summary.target_id)
@@ -112,7 +103,7 @@ impl SeoService {
                 )
                 .await?;
             let explicit = explicit_by_target.remove(&summary.target_id);
-            let meta = resolve_batched_bulk_meta(
+            let projection = resolve_bulk_read_projection(
                 tenant,
                 filter.target_kind.clone(),
                 summary.target_id,
@@ -123,7 +114,7 @@ impl SeoService {
             )
             .ok_or(SeoError::NotFound)?;
 
-            if filter.source != SeoBulkSource::Any && filter.source != meta.source {
+            if filter.source != SeoBulkSource::Any && filter.source != projection.source {
                 continue;
             }
             if let Some(query) = filter.query.as_deref() {
@@ -136,36 +127,64 @@ impl SeoService {
                     continue;
                 }
             }
-            rows.push(BatchedBulkRow { summary, meta });
+            rows.push(BulkReadRow {
+                target_id: summary.target_id,
+                label: summary.label,
+                route: summary.route,
+                projection,
+            });
         }
 
+        Ok(rows)
+    }
+
+    pub(super) async fn list_bulk_items_batched(
+        &self,
+        tenant: &TenantContext,
+        input: SeoBulkListInput,
+    ) -> SeoResult<SeoBulkPage> {
+        let page = input.page.max(1);
+        let per_page = input.per_page.clamp(1, MAX_BULK_PAGE_SIZE);
+        let filter = BulkReadFilter {
+            target_kind: input.target_kind,
+            locale: super::normalize_effective_locale(
+                input.locale.as_str(),
+                tenant.default_locale.as_str(),
+            )?,
+            query: input
+                .query
+                .map(|value| value.trim().to_ascii_lowercase())
+                .filter(|value| !value.is_empty()),
+            source: input.source.unwrap_or(SeoBulkSource::Any),
+        };
+        let rows = self.collect_bulk_read_rows(tenant, &filter).await?;
         let total = rows.len() as i32;
-        let offset = ((filter.page - 1) * filter.per_page) as usize;
+        let offset = ((page - 1) * per_page) as usize;
         let items = rows
             .into_iter()
             .skip(offset)
-            .take(filter.per_page as usize)
+            .take(per_page as usize)
             .map(|row| SeoBulkItem {
                 target_kind: filter.target_kind.clone(),
-                target_id: row.summary.target_id,
+                target_id: row.target_id,
                 locale: filter.locale.clone(),
-                effective_locale: row.meta.effective_locale,
-                label: row.summary.label,
-                route: row.summary.route,
-                source: row.meta.source,
-                title: row.meta.title,
-                description: row.meta.description,
-                canonical_url: row.meta.canonical_url,
-                noindex: row.meta.noindex,
-                nofollow: row.meta.nofollow,
+                effective_locale: row.projection.effective_locale,
+                label: row.label,
+                route: row.route,
+                source: row.projection.source,
+                title: row.projection.title,
+                description: row.projection.description,
+                canonical_url: row.projection.canonical_url,
+                noindex: row.projection.noindex,
+                nofollow: row.projection.nofollow,
             })
             .collect();
 
         Ok(SeoBulkPage {
             items,
             total,
-            page: filter.page,
-            per_page: filter.per_page,
+            page,
+            per_page,
         })
     }
 
@@ -216,34 +235,8 @@ impl SeoService {
     }
 }
 
-fn normalize_batched_bulk_list_input(
-    input: SeoBulkListInput,
-    fallback_locale: &str,
-) -> SeoResult<BatchedBulkListFilter> {
-    Ok(BatchedBulkListFilter {
-        target_kind: input.target_kind,
-        locale: super::super::normalize_effective_locale(input.locale.as_str(), fallback_locale)?,
-        query: input
-            .query
-            .map(|value| value.trim().to_ascii_lowercase())
-            .filter(|value| !value.is_empty()),
-        source: input.source.unwrap_or(SeoBulkSource::Any),
-        page: input.page.max(1),
-        per_page: input.per_page.clamp(1, MAX_BULK_PAGE_SIZE),
-    })
-}
-
-fn empty_bulk_page(filter: &BatchedBulkListFilter) -> SeoBulkPage {
-    SeoBulkPage {
-        items: Vec::new(),
-        total: 0,
-        page: filter.page,
-        per_page: filter.per_page,
-    }
-}
-
 #[allow(clippy::too_many_arguments)]
-fn resolve_batched_bulk_meta(
+fn resolve_bulk_read_projection(
     tenant: &TenantContext,
     target_kind: SeoTargetSlug,
     target_id: Uuid,
@@ -251,7 +244,7 @@ fn resolve_batched_bulk_meta(
     explicit: Option<LoadedMeta>,
     state: Option<TargetState>,
     settings: &SeoModuleSettings,
-) -> Option<BatchedBulkMeta> {
+) -> Option<BulkReadProjection> {
     match (explicit, state) {
         (Some(explicit), Some(state)) => {
             let resolved = resolve_by_locale_with_fallback(
@@ -261,7 +254,7 @@ fn resolve_batched_bulk_meta(
                 |item| item.locale.as_str(),
             );
             let translation = resolved.item.cloned();
-            Some(BatchedBulkMeta {
+            Some(BulkReadProjection {
                 effective_locale: resolved.effective_locale,
                 source: SeoBulkSource::Explicit,
                 title: translation
@@ -272,7 +265,20 @@ fn resolve_batched_bulk_meta(
                     .as_ref()
                     .and_then(|item| trimmed_option(item.description.clone()))
                     .or(state.description),
+                keywords: translation
+                    .as_ref()
+                    .and_then(|item| trimmed_option(item.keywords.clone())),
                 canonical_url: explicit.meta.canonical_url,
+                og_title: translation
+                    .as_ref()
+                    .and_then(|item| trimmed_option(item.og_title.clone())),
+                og_description: translation
+                    .as_ref()
+                    .and_then(|item| trimmed_option(item.og_description.clone())),
+                og_image: translation
+                    .as_ref()
+                    .and_then(|item| trimmed_option(item.og_image.clone())),
+                structured_data: explicit.meta.structured_data,
                 noindex: explicit.meta.no_index,
                 nofollow: explicit.meta.no_follow,
             })
@@ -285,14 +291,27 @@ fn resolve_batched_bulk_meta(
                 |item| item.locale.as_str(),
             );
             let translation = resolved.item.cloned();
-            Some(BatchedBulkMeta {
+            Some(BulkReadProjection {
                 effective_locale: resolved.effective_locale,
                 source: SeoBulkSource::Explicit,
                 title: translation.as_ref().and_then(|item| item.title.clone()),
                 description: translation
                     .as_ref()
                     .and_then(|item| item.description.clone()),
+                keywords: translation
+                    .as_ref()
+                    .and_then(|item| item.keywords.clone()),
                 canonical_url: explicit.meta.canonical_url,
+                og_title: translation
+                    .as_ref()
+                    .and_then(|item| item.og_title.clone()),
+                og_description: translation
+                    .as_ref()
+                    .and_then(|item| item.og_description.clone()),
+                og_image: translation
+                    .as_ref()
+                    .and_then(|item| item.og_image.clone()),
+                structured_data: explicit.meta.structured_data,
                 noindex: explicit.meta.no_index,
                 nofollow: explicit.meta.no_follow,
             })
@@ -314,7 +333,7 @@ fn resolve_batched_bulk_meta(
                 || generated.robots.is_some()
                 || generated.twitter_title.is_some()
                 || generated.twitter_description.is_some();
-            Some(BatchedBulkMeta {
+            Some(BulkReadProjection {
                 effective_locale: state.effective_locale,
                 source: if generated_source {
                     SeoBulkSource::Generated
@@ -323,7 +342,14 @@ fn resolve_batched_bulk_meta(
                 },
                 title: generated.title.or(Some(state.title)),
                 description: generated.description.or(state.description),
+                keywords: generated.keywords,
                 canonical_url: generated.canonical_url,
+                og_title: generated.og_title.or(state.open_graph.title.clone()),
+                og_description: generated
+                    .og_description
+                    .or(state.open_graph.description.clone()),
+                og_image: first_open_graph_image_url(&state.open_graph),
+                structured_data: Some(state.structured_data),
                 noindex: false,
                 nofollow: false,
             })
@@ -372,59 +398,11 @@ mod tests {
         }
     }
 
-    fn explicit(target_kind: &SeoTargetSlug, target_id: Uuid) -> LoadedMeta {
-        let meta_id = Uuid::new_v4();
-        LoadedMeta {
-            meta: seo_meta::Model {
-                id: meta_id,
-                tenant_id: tenant().id,
-                target_type: target_kind.as_str().to_string(),
-                target_id,
-                no_index: true,
-                no_follow: false,
-                canonical_url: Some("/explicit".to_string()),
-                structured_data: Some(json!({"@type": "WebPage"})),
-            },
-            translations: vec![meta_translation::Model {
-                id: Uuid::new_v4(),
-                meta_id,
-                locale: "en-US".to_string(),
-                title: Some("Explicit title".to_string()),
-                description: Some("Explicit description".to_string()),
-                keywords: None,
-                og_title: None,
-                og_description: None,
-                og_image: None,
-            }],
-        }
-    }
-
     #[test]
-    fn explicit_projection_preserves_explicit_fields() {
+    fn fallback_projection_preserves_full_export_fields() {
         let target_kind = SeoTargetSlug::new("page").expect("valid target kind");
         let target_id = Uuid::new_v4();
-        let resolved = resolve_batched_bulk_meta(
-            &tenant(),
-            target_kind.clone(),
-            target_id,
-            "en-US",
-            Some(explicit(&target_kind, target_id)),
-            Some(state(target_kind, target_id)),
-            &SeoModuleSettings::default(),
-        )
-        .expect("projection");
-
-        assert_eq!(resolved.source, SeoBulkSource::Explicit);
-        assert_eq!(resolved.title.as_deref(), Some("Explicit title"));
-        assert_eq!(resolved.canonical_url.as_deref(), Some("/explicit"));
-        assert!(resolved.noindex);
-    }
-
-    #[test]
-    fn fallback_projection_uses_target_fields_without_templates() {
-        let target_kind = SeoTargetSlug::new("page").expect("valid target kind");
-        let target_id = Uuid::new_v4();
-        let resolved = resolve_batched_bulk_meta(
+        let projection = resolve_bulk_read_projection(
             &tenant(),
             target_kind.clone(),
             target_id,
@@ -435,30 +413,8 @@ mod tests {
         )
         .expect("projection");
 
-        assert_eq!(resolved.source, SeoBulkSource::Fallback);
-        assert_eq!(resolved.title.as_deref(), Some("Fallback title"));
-        assert_eq!(resolved.description.as_deref(), Some("Fallback description"));
-        assert!(!resolved.noindex);
-    }
-
-    #[test]
-    fn normalization_bounds_page_and_canonicalizes_query() {
-        let filter = normalize_batched_bulk_list_input(
-            SeoBulkListInput {
-                target_kind: SeoTargetSlug::new("page").expect("valid target kind"),
-                locale: "en-us".to_string(),
-                query: Some("  Sale  ".to_string()),
-                source: None,
-                page: 0,
-                per_page: 999,
-            },
-            "en-US",
-        )
-        .expect("filter");
-
-        assert_eq!(filter.locale, "en-US");
-        assert_eq!(filter.query.as_deref(), Some("sale"));
-        assert_eq!(filter.page, 1);
-        assert_eq!(filter.per_page, MAX_BULK_PAGE_SIZE);
+        assert_eq!(projection.source, SeoBulkSource::Fallback);
+        assert_eq!(projection.title.as_deref(), Some("Fallback title"));
+        assert_eq!(projection.structured_data, Some(json!({"@type": "WebPage"})));
     }
 }
